Integreated libbullseye compiled from source
I have modified and added scripts to compile and link bullseye
Here libbullsye uses the llvm10 that is built for pluto and
statically links with cplex that is being used.

At the current stage libbullseye clones and uses it's own
isl and pet, and is not dependent on pluto to provide the same.